[ feat ] 8 : 후원 프로젝트 수정 기능 구현

+++ backend/src/main/java/com/funding/backend/domain/donation/repository/DonationRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DonationRepository extends JpaRepository<Donation,Integer> {
+public interface DonationRepository extends JpaRepository<Donation, Long> {
 
   Optional<Donation> findByProject(Project project);
 }
